Add files via upload
Encountered a problem earlier where I downloaded a the file to continue working, and the file was saved as VBA_Challenge(1).  This mistake created multiple versions of the excel file. Just now, I downloaded the file to verify it was working correctly, I realized the most recent upload was a less developed version.  Uploading the correct version.